cl_dll: 64-bit client for the Xash3D-FWGS-hosted client (static openvgui, no SDL2)
Port of goldsrc-net/mod-ts 94c0bba + 8f3b248 (branch modernize-build-sdl2-vgui) onto 64bit-demo:
- 64-bit targets static-link third_party/openvgui (FWGS 1b89d19) and the vgui_support shim
  (goldsrc-net/vgui_support@emscripten b6d5e1d, INTERNAL_VGUI_SUPPORT) so the client exports
  InitVGUISupportAPI and carries no vgui.so / libSDL2 DT_NEEDED; 32-bit keeps Valve's vgui.so + SDL2.
- Input comes from the engine push callbacks IN_ClientLookEvent / IN_ClientMoveEvent (inputw32.cpp),
  SDL confined to i386 Linux; SDL_ShowSimpleMessageBox -> Con_Printf.
- particleman/entity.cpp portability for gcc-8 (buildchain image).
- HALFLIFE_ALLOW_VALVE_GAMEDIR (64-bit): skip the UTIL_IsValveGameDirectory refusal — the Godot host
  runs this SDK as the Valve base game itself.
- stub_AllocString string pool for client-side weapon prediction: 64-bit MAKE_STRING is
  ALLOC_STRING and the client engine API has no AllocString (CGlock::Spawn crashed on NULL).
Built with: make build-halflife-updated-client-amd64 (hlds64-buildchain). Loaded by the vendored
Xash client half in goldsrc.net/godot (Phase 3).

+++ cl_dll/entity.cpp
@@ -514,8 +514,9 @@ void DLLEXPORT HUD_TempEntUpdate(
 				s = sin(pTemp->entity.baseline.origin[2] + fastFreq);
 				c = cos(pTemp->entity.baseline.origin[2] + fastFreq);
 
-				pTemp->entity.origin[0] += pTemp->entity.baseline.origin[0] * frametime + 8 * sin(client_time * 20 + (int)pTemp);
-				pTemp->entity.origin[1] += pTemp->entity.baseline.origin[1] * frametime + 4 * sin(client_time * 30 + (int)pTemp);
+				// (int)(intptr_t) — pTemp is just a per-entity phase seed; truncation is fine (64-bit safe).
+				pTemp->entity.origin[0] += pTemp->entity.baseline.origin[0] * frametime + 8 * sin(client_time * 20 + (int)(intptr_t)pTemp);
+				pTemp->entity.origin[1] += pTemp->entity.baseline.origin[1] * frametime + 4 * sin(client_time * 30 + (int)(intptr_t)pTemp);
 				pTemp->entity.origin[2] += pTemp->entity.baseline.origin[2] * frametime;
 			}
